bspline,gui: widen the engraving tick accounting to 64 bits
The firmware target is pico-plus2 / RP2350, where uint is 32 BITS, and a
plate's tick count no longer fits there. Measured, as a fraction of
MaxUint32:

                8.0mm/s x1  1.0mm/s x1  1.0mm/s x8
    CONSTPROOF!     0.045x      0.155x      1.216x   (453 min)
    TEXTPROOF!      0.042x      0.154x      1.198x   (446 min)
    BOTHPROOF!      0.038x      0.130x      1.018x   (379 min)

Speed alone tops out at 0.155x -- 6.4x of headroom, which is why the
speed slice needed nothing. Passes multiply that by 7.8x (not the full 8:
travel between glyphs is cut once however many times each glyph is
struck) and carry all THREE proof patterns over in one step.

On device the CONSTPROOF! duration wrapped to ~926.7M, so the engrave
screen showed "80:26" for a seven-and-a-half-hour job and then
unsigned-underflowed partway through. Engraving itself was never
affected -- completion is signalled by the job goroutine's error channel,
not by the tick count -- but a countdown that lies by a factor of five
about a seven-hour job is not acceptable.

Widen rather than clamp, because the wrap is in the ACCUMULATION and not
only in the display: bspline.Measure sums per-knot T into the total, so
on a 32-bit target the number is already wrong before any display code
sees it, and a clamp would be formatting a value with no relation to the
job. Knot.T is a genuinely small per-knot count and stays uint.

The progress channel is widened too, deliberately: reportProgress folds
each unread report into the pending one, so an uncollected channel
carries a running total rather than a delta.

Second defect fixed in the same expression: `s.duration - st.Completed`
is unsigned and Completed can legitimately EXCEED duration, because a
resumed job also cuts synthesised catch-up motion (SafePointer.Resume)
that the plan never counted. Unsigned, that wraps to near 2^64 and prints
a ten-digit hour count. The countdown now saturates at 0:00.

The four lines move into engraveRemaining() so the arithmetic can be
asserted at durations no test can reach by actually engraving.

The type pin is by REFLECTION and not by arithmetic: uint and uint64 are
the same type on the amd64 host the tests run on, so no value assertion
could tell them apart -- which is exactly why this shipped. Reflection
distinguishes them on every platform.

cmd/plateview widens to match; its own comment says a preview that
estimated differently from the machine would be a second answer to a
question with one, and that applies to the width as much as the formula.

Firmware size: 2,532,352 -> 2,533,376 bytes, +512 bytes of flash payload,
+0.04%. The 64-bit adds are in Measure (once per knot at plan time) and
in Status/draw (twice a second); nothing on the stepper path changed.

+++ cmd/plateview/main.go
@@ -130,8 +130,12 @@ func sizeLabel(p gui.Preview) string {
 // the device's own progress screen does -- divide by TicksPerSecond, round up
 // (gui.go's engraveRunning case). A preview that estimated differently from
 // the machine would be a second answer to a question with one.
-func engraveDuration(ticks uint) string {
-	tps := uint(engraverParams.TicksPerSecond)
+// The tick count is uint64 for the reason the firmware's is: the widest real
+// plate is past MaxUint32 (see bspline.Attributes.Duration). This tool runs on
+// a 64-bit host where uint would work, but a preview whose arithmetic differed
+// in WIDTH from the machine's is the second answer this comment warns about.
+func engraveDuration(ticks uint64) string {
+	tps := uint64(engraverParams.TicksPerSecond)
 	secs := (ticks + tps - 1) / tps
 	return fmt.Sprintf("%dm%02ds", secs/60, secs%60)
 }
